requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.1
django-bootstrap4==1.1.1
idna==2.8
maxminddb==1.5.2
maxminddb-geolite2==2018.703
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
python-geoip==1.2
python-geoip-geolite2==2015.303
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7